+++ apps/tabler-app/src/queries/Member/MemberDetailsFragment.ts
@@ -18,6 +18,10 @@ export const MemberDetailsFragment = gql`
             value
         }
 
+        association {
+            isocode
+        }
+
         phonenumbers {
             type
             value