+++ services/tabler-world-api/src/graphql/schema/Structure.ts
@@ -19,6 +19,7 @@ export const Structure = gql`
 
         logo: String
         flag: String
+        isocode: String!
 
         name: String!
         shortname: String!