+++ apps/tabler-app/assets/docs/releasenotes.md
@@ -5,6 +5,7 @@
 - TABLER.APP now speaks Finnish, thanks to [Teemu Keski-Valkama](tablerworld:/member?id=121983)
 - TABLER.APP now speaks Icelandic, thanks to [Daníel Sigurður Eðvaldsson](tablerworld:/member?id=128546)
 - TABLER.APP now speaks Italien, thanks to Franco Cavalli
+- We now format all phone numbers in international format
 
 ## Fixes
 - Fixed an issue with cameral roll permissions on Android